+++ test/test_GalKin/test_gom.py
@@ -31,8 +31,8 @@ class TestGOM(object):
 
         psf_fwhm = 1.  # Gaussian FWHM psf
         kwargs_cosmo = {'d_d': 1000, 'd_s': 1500, 'd_ds': 800}
-        kwargs_numerics = {'interpol_grid_num': 100, 'log_integration': True,
-                           'max_integrate': 100, 'min_integrate': 0.001}
+        kwargs_numerics = {'interpol_grid_num': 2000, 'log_integration': True,
+                           'max_integrate': 100, 'min_integrate': 0.00001}
 
         # anisotropy profile
         anisotropy_type = 'OM'
@@ -61,4 +61,4 @@ class TestGOM(object):
         sigma_v_gom = galkin_gom.dispersion(kwargs_profile, kwargs_light, kwargs_anisotropy, sampling_number=5000)
         # warning: this tests does not work to this precision for every random seed. To increase precision, increase
         # sampling_number
-        npt.assert_almost_equal(sigma_v_gom, sigma_v_om, decimal=0)
+        npt.assert_almost_equal(sigma_v_gom / sigma_v_om, 1, decimal=2)